feat: moved pdf to separate component
