--- apps/gauzy/src/app/pages/invoices/invoice-download/invoice-download-mutation.component.scss
@@ -1,8 +1,0 @@
-.main {
-  height: 90vh;
-}
-
-.pdfDoc {
-  height: 100%;
-  width: 100%;
-}
