--- apps/gauzy/src/app/pages/invoices/invoice-email/invoice-email-mutation.component.scss
@@ -1,12 +1,0 @@
-.main {
-  height: 90vh;
-}
-
-.pdfDoc {
-  height: 80%;
-  width: 100%;
-}
-
-.invoice-email-body {
-  padding-bottom: 0;
-}
