+++ apps/gauzy/src/app/pages/invoices/invoice-send/invoice-send-mutation.component.ts
@@ -2,28 +2,14 @@ import { Component, OnInit } from '@angular/core';
 import { NbDialogRef, NbToastrService } from '@nebular/theme';
 import { TranslateService } from '@ngx-translate/core';
 import { TranslationBaseComponent } from '../../../@shared/language-base/translation-base.component';
-import {
-	IInvoice,
-	ITag,
-	InvoiceStatusTypesEnum,
-	InvoiceTypeEnum
-} from '@gauzy/models';
+import { IInvoice, ITag, InvoiceStatusTypesEnum } from '@gauzy/models';
 import { InvoicesService } from '../../../@core/services/invoices.service';
 import { Store } from '../../../@core/services/store.service';
 import { InvoiceEstimateHistoryService } from '../../../@core/services/invoice-estimate-history.service';
-import { EmployeesService } from '../../../@core/services';
-import { OrganizationProjectsService } from '../../../@core/services/organization-projects.service';
-import { TasksService } from '../../../@core/services/tasks.service';
-import { ProductService } from '../../../@core/services/product.service';
-import { ExpensesService } from '../../../@core/services/expenses.service';
-import { generatePdf } from '../../../@shared/invoice/generate-pdf';
-import * as pdfMake from 'pdfmake/build/pdfmake';
-import * as pdfFonts from 'pdfmake/build/vfs_fonts';
 
 @Component({
 	selector: 'ga-invoice-send',
-	templateUrl: './invoice-send-mutation.component.html',
-	styleUrls: ['./invoice-send-mutation.component.scss']
+	templateUrl: './invoice-send-mutation.component.html'
 })
 export class InvoiceSendMutationComponent
 	extends TranslationBaseComponent
@@ -32,7 +18,6 @@ export class InvoiceSendMutationComponent
 	alreadySent = false;
 	tags: ITag[];
 	isEstimate: boolean;
-	docDefinition: any;
 
 	constructor(
 		protected dialogRef: NbDialogRef<InvoiceSendMutationComponent>,
@@ -40,12 +25,7 @@ export class InvoiceSendMutationComponent
 		readonly translateService: TranslateService,
 		private toastrService: NbToastrService,
 		private invoiceEstimateHistoryService: InvoiceEstimateHistoryService,
-		private store: Store,
-		private employeeService: EmployeesService,
-		private projectService: OrganizationProjectsService,
-		private taskService: TasksService,
-		private productService: ProductService,
-		private expensesService: ExpensesService
+		private store: Store
 	) {
 		super(translateService);
 	}
@@ -54,50 +34,6 @@ export class InvoiceSendMutationComponent
 		if (this.invoice.sentTo) {
 			this.alreadySent = true;
 		}
-		this.loadPdf();
-	}
-
-	async loadPdf() {
-		pdfMake.vfs = pdfFonts.pdfMake.vfs;
-		let docDefinition;
-		let service;
-
-		switch (this.invoice.invoiceType) {
-			case InvoiceTypeEnum.BY_EMPLOYEE_HOURS:
-				service = this.employeeService;
-				break;
-			case InvoiceTypeEnum.BY_PROJECT_HOURS:
-				service = this.projectService;
-				break;
-			case InvoiceTypeEnum.BY_TASK_HOURS:
-				service = this.taskService;
-				break;
-			case InvoiceTypeEnum.BY_PRODUCTS:
-				service = this.productService;
-				break;
-			case InvoiceTypeEnum.BY_EXPENSES:
-				service = this.expensesService;
-				break;
-			default:
-				break;
-		}
-
-		docDefinition = await generatePdf(
-			this.invoice,
-			this.invoice.fromOrganization,
-			this.invoice.toContact,
-			service
-		);
-
-		this.docDefinition = docDefinition;
-
-		const pdfDocGenerator = pdfMake.createPdf(docDefinition);
-		pdfDocGenerator.getDataUrl((dataUrl) => {
-			const iframe = document.querySelector(
-				'#iframe'
-			) as HTMLIFrameElement;
-			iframe.src = dataUrl;
-		});
 	}
 
 	async closeDialog() {
